Improvements over Rectangle struct (#8)
* fix FromLTRB signature

* Fixed misspell

* Add .editorconfig to solution file

---------

+++ src/Common/Rectangle.cs
@@ -70,7 +70,7 @@ public struct Rectangle : IEquatable<Rectangle>
 	public readonly bool Equals(Rectangle other) => m_rect == other.m_rect;
 
 	/// <summary>
-	/// Tests whether <paramref name="obj"/> is a <see cref='Rectangle'/> with the same locaation
+	/// Tests whether <paramref name="obj"/> is a <see cref='Rectangle'/> with the same location
 	/// and size as this Rectangle.
 	/// </summary>
 	public override readonly bool Equals(object obj) => m_rect.Equals(obj);
@@ -198,7 +198,7 @@ public struct Rectangle : IEquatable<Rectangle>
 	/// <summary>
 	/// Creates a new <see cref='Rectangle'/> with the specified location and size.
 	/// </summary>
-	public static Rectangle FromLTRB(int left, int top, int right, int bottom) => new(left, top, unchecked(right - left), unchecked(bottom - top));
+	public static Rectangle FromLTRB(float left, float top, float right, float bottom) => new(left, top, unchecked(right - left), unchecked(bottom - top));
 
 	#endregion
